Changed definition of linear combinations again
Linear combinations now require the list to have unique vectors.  This
complicated some of the earlier proofs, but it should make some things
much easier in the future.

Another addition related to this was the addition of
linear_combination_of, which checks if a vector can be written as a
linear combination of vectors in some set.