Partly made injection from vectors to functionals
It's partly made because I've realized that the whole point was to
define something close enough to the tensor algebra, but now I realized
that this construction of the tensor algebra requires division, when the
standard construction does not.  Thus, I gave up partway through because
I'm going to need a new construction.