Completely revamped equivalence.v
There were four major changes that happened:
 - The development of the theory of operations on quotients was
   overhauled and now has a more categorical flavor
 - binary_rop and all self_ops were removed
 - The definitions of to_equiv_type, unary_op, and binary_op were made
   opaque
 - Several names were changed to be shorter (such as to_equiv_type ->
   to_equiv) or more consistent with the rest of the project (such as
   equiv_binary_op -> binary_op_eq)

binary_rop was removed because you can get away with using an unary_op:
Given a function A → B → C, we can convert it to a function A → B/~ → C
by fixing a value a : A and then using unary_op to convert B → C to B/~
→ C.

The self_ops were removed because you can just consider a function A → A
as a function A → A/~, which you can then convert to A/~ → A/~.  The
theorems unary_self_wd and binary_self_wd were created to make this
process easier.

The definitions were made opaque for complicated reasons that are
described in a code comment in equivalence.v.