Made Cauchy reals not depend on card, using them
By "using them", I mean that the reals are defined using the Cauchy
sequence construction, not the Dedekind cut construction.  Now that the
Cauchy sequence construction doesn't depend on the cardinals, they are
the fastest to compile of all three constructions, both from a fresh
state and from just recompiling the constructions themselves.